+++ src/main/java/com/ketheroth/uncrafter/common/inventory/container/UncrafterContainer.java
@@ -1,13 +1,16 @@
 package com.ketheroth.uncrafter.common.inventory.container;
 
 import com.ketheroth.uncrafter.common.config.Configuration;
+import com.ketheroth.uncrafter.common.tileentity.UncrafterTileEntity;
 import com.ketheroth.uncrafter.core.registry.UncrafterContainerTypes;
 import mcp.MethodsReturnNonnullByDefault;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.entity.player.PlayerInventory;
 import net.minecraft.entity.player.ServerPlayerEntity;
+import net.minecraft.inventory.container.ClickType;
 import net.minecraft.inventory.container.Container;
 import net.minecraft.inventory.container.Slot;
+import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
 import net.minecraft.item.crafting.RecipeManager;
 import net.minecraft.util.Tuple;
@@ -24,40 +27,34 @@ import java.util.List;
 
 @ParametersAreNonnullByDefault
 @MethodsReturnNonnullByDefault
-public class UncrafterContainer extends Container implements IUncrafterContainer {
+public class UncrafterContainer extends Container {
 
 	private final BlockPos pos;
 	private final PlayerEntity player;
+	private final UncrafterTileEntity uncrafter;
 	private final IItemHandler playerInventory;
 	private final OutputHandler outputItems;
 	private final InputHandler inputItems;
-	private Tuple<ItemStack, List<ItemStack>> cache = new Tuple<>(ItemStack.EMPTY, new ArrayList<>());
 
 	public UncrafterContainer(int windowId, PlayerInventory playerInventory, PlayerEntity player, BlockPos pos) {
 		super(UncrafterContainerTypes.UNCRAFTER_CONTAINER.get(), windowId);
 		this.pos = pos;
 		this.player = player;
+		this.uncrafter = ((UncrafterTileEntity) player.level.getBlockEntity(this.pos));
 		this.playerInventory = new InvWrapper(playerInventory);
-		this.outputItems = new OutputHandler(9, this);
-		this.inputItems = new InputHandler(1, this);
+		this.outputItems = this.uncrafter.getOutput();
+		this.inputItems = this.uncrafter.getInput();
 
 		//layout input inventory
 		this.addSlot(new SlotItemHandler(this.inputItems, 0, 35, 35) {
 			@Override
 			public boolean mayPlace(@Nonnull ItemStack stack) {
-				if (stack.getItem().getRegistryName() == null) {
-					return false;
-				}
-				String name = stack.getItem().getRegistryName().toString();
-				if (Configuration.WHITELIST.get().isEmpty()) {
-					return !Configuration.BLACKLIST.get().contains(name) && !Configuration.IMC_BLACKLIST.contains(name);
-				}
-				return Configuration.WHITELIST.get().contains(name);
+				return Configuration.isValidItem(stack.getItem());
 			}
 
 			@Override
 			public boolean mayPickup(PlayerEntity playerIn) {
-				return !UncrafterContainer.this.isInputLocked() && super.mayPickup(playerIn);
+				return !UncrafterContainer.this.uncrafter.isInputLocked() && super.mayPickup(playerIn);
 			}
 		});
 		//layout output inventory
@@ -88,17 +85,25 @@ public class UncrafterContainer extends Container implements IUncrafterContainer
 	}
 
 	@Override
-	public void removed(PlayerEntity player) {
-		if (player instanceof ServerPlayerEntity) {
-			ItemStack itemstack = this.inputItems.extractItem(0, 64, false);
-			if (!itemstack.isEmpty()) {
-				if (player.isAlive() && !((ServerPlayerEntity) player).hasDisconnected()) {
-					player.inventory.placeItemBackInInventory(this.player.level, itemstack);
-				} else {
-					player.drop(itemstack, false);
+	public ItemStack clicked(int index, int quickCraft, ClickType clickType, PlayerEntity player) {
+		if (1 <= index && index < 10) {
+			int maxExtract = this.uncrafter.isAdvanced() ? Configuration.ADVANCED_EXTRACT_AMOUNT.get() : Configuration.EXTRACT_AMOUNT.get();
+			if (this.outputItems.getStackInSlot(index - 1).isEmpty()) {
+				int n = 0;
+				for (int i = 0; i < 9; i++) {
+					if (this.uncrafter.selected[i]) {
+						n++;
+					}
 				}
+				if (this.uncrafter.selected[index - 1]) {
+					this.uncrafter.selected[index - 1] = false;
+				} else {
+					this.uncrafter.selected[index - 1] = n < maxExtract;
+				}
+				this.uncrafter.setChanged();
 			}
 		}
+		return super.clicked(index, quickCraft, clickType, player);
 	}
 
 	@Override
@@ -118,13 +123,17 @@ public class UncrafterContainer extends Container implements IUncrafterContainer
 					}
 				}
 			} else if (1 <= index && index < 10) {// shift-click in output slots
-				// nothing happens
+				if (!this.moveItemStackTo(slotStack, 10, 46, true)) {
+					return ItemStack.EMPTY;
+				}
+				this.outputItems.extractItem(index - 1, 1, false);
+				return ItemStack.EMPTY;
 			} else if (index < 46) {// shift-click in inventory slots
 				if (!slots.get(0).hasItem() || slotStack.sameItem(slots.get(0).getItem())) {//shift click from inventory
 					if (!this.moveItemStackTo(slotStack, 0, 1, false)) {
 						return ItemStack.EMPTY;
 					}
-					inputItems.fillOutputSlots(itemstack);
+					this.inputItems.fillOutputSlots(itemstack);
 				}
 			}
 
@@ -141,55 +150,12 @@ public class UncrafterContainer extends Container implements IUncrafterContainer
 		return itemstack;
 	}
 
-	@Override
-	public boolean canTakeItemForPickAll(ItemStack pStack, Slot pSlot) {
-		return false;
-	}
-
-	@Override
 	public boolean isInputLocked() {
-		return this.outputItems.isExtracting();
+		return this.uncrafter.isInputLocked();
 	}
 
-	@Override
-	public InputHandler getInputHandler() {
-		return this.inputItems;
-	}
-
-	@Override
-	public OutputHandler getOutputHandler() {
-		return this.outputItems;
-	}
-
-	@Nullable
-	@Override
-	public EnchantmentHandler getEnchantmentHandler() {
-		return null;
-	}
-
-	@Override
-	public Tuple<ItemStack, List<ItemStack>> getCache() {
-		return this.cache;
-	}
-
-	@Override
-	public void setCache(List<ItemStack> b) {
-		this.cache = new Tuple<>(this.cache.getA(), b);
-	}
-
-	@Override
-	public void setCache(ItemStack a, List<ItemStack> b) {
-		this.cache = new Tuple<>(a, b);
-	}
-
-	@Override
-	public RecipeManager getRecipeManager() {
-		return this.player.level.getRecipeManager();
-	}
-
-	@Override
-	public boolean isAdvanced() {
-		return false;
+	public boolean[] selectedIndexes() {
+		return this.uncrafter.selected;
 	}
 
 }